# API_DOCUMENTATION.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Finance Data Processing API",
                 "description":  "Backend API for Finance Dashboard with Role-Based Access Control",
                 "contact":  {
                                 "name":  "Md Shahbaz Imam",
                                 "email":  "mdshahbazimam02@gmail.com"
                             },
                 "version":  "1.0.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Default Server"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
...
    "paths":  {
                  "/api/transactions/{id}":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "Transactions"
                                                                      ],
                                                             "summary":  "Get transaction by ID",
                                                             "operationId":  "getById",
                                                             "parameters":  [
                                                                                {
                                                                                    "name":  "id",
                                                                                    "in":  "path",
                                                                                    "required":  true,
                                                                                    "schema":  {
                                                                                                   "type":  "integer",
                                                                                                   "format":  "int64"
                                                                                               }
                                                                                }
                                                                            ],
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ApiResponseTransactionResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       }
                                                                           },
                                                             "security":  [
                                                                              {
                                                                                  "bearerAuth":  [

                                                                                                 ]
                                                                              }
                                                                          ]
                                                         },
                                                 "put":  {
                                                             "tags":  [
                                                                          "Transactions"
                                                                      ],
                                                             "summary":  "Update a transaction (ADMIN only)",
                                                             "operationId":  "update",
                                                             "parameters":  [
                                                                                {
                                                                                    "name":  "id",
                                                                                    "in":  "path",
                                                                                    "required":  true,
                                                                                    "schema":  {
                                                                                                   "type":  "integer",
                                                                                                   "format":  "int64"
                                                                                               }
                                                                                }
                                                                            ],
                                                             "requestBody":  {
                                                                                 "content":  {
                                                                                                 "application/json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/TransactionRequest"
                                                                                                                                     }
                                                                                                                      }
                                                                                             },
                                                                                 "required":  true
                                                                             },
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ApiResponseTransactionResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       }
                                                                           },
                                                             "security":  [
                                                                              {
                                                                                  "bearerAuth":  [

                                                                                                 ]
                                                                              }
                                                                          ]
                                                         },
                                                 "delete":  {
                                                                "tags":  [
                                                                             "Transactions"
                                                                         ],
                                                                "summary":  "Soft delete a transaction (ADMIN only)",
                                                                "operationId":  "delete",
                                                                "parameters":  [
                                                                                   {
                                                                                       "name":  "id",
                                                                                       "in":  "path",
                                                                                       "required":  true,
                                                                                       "schema":  {
                                                                                                      "type":  "integer",
                                                                                                      "format":  "int64"
                                                                                                  }
                                                                                   }
                                                                               ],
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "OK",
                                                                                              "content":  {
                                                                                                              "*/*":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/ApiResponseVoid"
                                                                                                                                     }
                                                                                                                      }
                                                                                                          }
                                                                                          }
                                                                              },
                                                                "security":  [
                                                                                 {
                                                                                     "bearerAuth":  [

                                                                                                    ]
                                                                                 }
                                                                             ]
                                                            }
                                             },
...
              },
    "components":  {
                       "schemas":  {
                                       "TransactionRequest":  {
                                                                  "required":  [
                                                                                   "amount",
                                                                                   "category",
                                                                                   "date",
                                                                                   "type"
                                                                               ],
                                                                  "type":  "object",
                                                                  "properties":  {
                                                                                     "amount":  {
                                                                                                    "minimum":  0.01,
                                                                                                    "exclusiveMinimum":  false,
                                                                                                    "type":  "number"
                                                                                                },
                                                                                     "type":  {
                                                                                                  "type":  "string",
                                                                                                  "enum":  [
                                                                                                               "INCOME",
                                                                                                               "EXPENSE"
                                                                                                           ]
                                                                                              },
                                                                                     "category":  {
                                                                                                      "maxLength":  100,
                                                                                                      "minLength":  0,
                                                                                                      "type":  "string"
                                                                                                  },
                                                                                     "date":  {
                                                                                                  "type":  "string",
                                                                                                  "format":  "date"
                                                                                              },
                                                                                     "notes":  {
                                                                                                   "maxLength":  500,
                                                                                                   "minLength":  0,
                                                                                                   "type":  "string"
                                                                                               }
                                                                                 }
                                                              },
                                       "ApiResponseTransactionResponse":  {
                                                                              "type":  "object",
                                                                              "properties":  {
                                                                                                 "success":  {
                                                                                                                 "type":  "boolean"
                                                                                                             },
                                                                                                 "message":  {
                                                                                                                 "type":  "string"
                                                                                                             },
                                                                                                 "data":  {
                                                                                                              "$ref":  "#/components/schemas/TransactionResponse"
                                                                                                          },
                                                                                                 "timestamp":  {
                                                                                                                   "type":  "string",
                                                                                                                   "format":  "date-time"
                                                                                                               }
                                                                                             }
                                                                          },
                                       "TransactionResponse":  {
                                                                   "type":  "object",
                                                                   "properties":  {
                                                                                      "id":  {
                                                                                                 "type":  "integer",
                                                                                                 "format":  "int64"
                                                                                             },
                                                                                      "amount":  {
                                                                                                     "type":  "number"
                                                                                                 },
                                                                                      "type":  {
                                                                                                   "type":  "string",
                                                                                                   "enum":  [
                                                                                                                "INCOME",
                                                                                                                "EXPENSE"
                                                                                                            ]
                                                                                               },
                                                                                      "category":  {
                                                                                                       "type":  "string"
                                                                                                   },
                                                                                      "date":  {
                                                                                                   "type":  "string",
                                                                                                   "format":  "date"
                                                                                               },
                                                                                      "notes":  {
                                                                                                    "type":  "string"
                                                                                                },
                                                                                      "createdByName":  {
                                                                                                            "type":  "string"
                                                                                                        },
                                                                                      "createdAt":  {
                                                                                                        "type":  "string",
                                                                                                        "format":  "date-time"
                                                                                                    },
                                                                                      "updatedAt":  {
                                                                                                        "type":  "string",
                                                                                                        "format":  "date-time"
                                                                                                    }
                                                                                  }
                                                               },
...
                                       "ApiResponseVoid":  {
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "success":  {
                                                                                                  "type":  "boolean"
                                                                                              },
                                                                                  "message":  {
                                                                                                  "type":  "string"
                                                                                              },
                                                                                  "data":  {
                                                                                               "type":  "object"
                                                                                           },
                                                                                  "timestamp":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "date-time"
                                                                                                }
                                                                              }
                                                           }
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "description":  "JWT Bearer Token. Format: Bearer \u003ctoken\u003e",
                                                                  "in":  "header",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}